refactor(gateway): extract pge_loop.py from gateway.py — split step 4/6
Schritt 4 von 6 im Gateway-Split-Plan — der größte Brocken.

What moved
- 3 PGE-orchestration helpers extracted into `gateway/pge_loop.py`:
    * `is_cu_plan(plan)` — pure helper, no `gw` parameter (was @staticmethod)
    * `run_pge_loop(gw, msg, session, wm, ...)` (~900 LOC) — Phase 3
      Plan→Gate→Execute orchestration with replan loop and stuck-detection
    * `handle_approvals(gw, steps, decisions, session, channel_name, *,
      ws_session_id=None)` — ORANGE approval flow, blocks until user
      decision arrives via the channel

`gateway.py`: 5 205 -> 4 249 lines (-18 % in this PR; cumulative -36 %
from the pre-split 6 628 LOC after PRs 1-4).

Critical preservation: stalled-counter and `t` shadowing
- `MAX_STALLED_MODEL_TURNS` and `advance_stalled_count()` STAY at module
  level in `gateway.py` because `tests/test_core/test_stalled_counter.py:5`
  imports them as `from cognithor.gateway.gateway import ...`. `pge_loop.py`
  reads them through that import — clean dependency direction.
- `_TOOL_STATUS_KEYS` and `_sanitize_broken_llm_output` likewise stay in
  `gateway.py` and are imported from there. Single source of truth.
- The `t in _CORRECTION_TRIGGERS` comprehension uses `t` as loop variable
  inside the same scope where `t` is imported from `cognithor.i18n`.
  Python's comprehension scope keeps them separate (verified via test run).

Lessons captured during dev (continuing from PRs 1-3)
- **Off-by-one in slice end**: my first slice `lines[5114:5199]` extracted
  the def-line of `_track_reddit_replies` (next method) along with
  `_handle_approvals`. Caught when import of gateway.py left an orphan
  function body without def. Fix: trimmed pge_loop.py + restored the
  `_track_reddit_replies` stub method in gateway.py with its body.
- Imports needed beyond script-detected ones: `hashlib`, `json as _json`,
  `Message`, `MessageRole`, `t` (i18n), plus internal-pkg imports
  (`_TOOL_STATUS_KEYS`, `_sanitize_broken_llm_output`).
- `_is_cu_plan` was `@staticmethod` — wrapper preserves the decorator;
  the extracted function dropped the decorator (it's a free function).

Verification
- `pytest tests/test_gateway/ tests/test_core/` -> **2 704 passed, 1 skipped**.
- `pytest tests/test_core/test_stalled_counter.py` (verifies the
  `MAX_STALLED_MODEL_TURNS` import path is preserved) -> green.
- `ruff check + format` -> All checks passed.
- `python -c "from cognithor.gateway.gateway import Gateway; from cognithor.gateway import pge_loop"` -> OK.

Plan reference: project_v0960_refactor_backlog.md "gateway.py split"
+ inline architect blueprint. Next PR: `message_handler.py` (PR 5/6) —
extracts `handle_message` itself plus `_resolve_agent_route` and
`_prepare_execution_context`.
